+++ packages/javascript/lib/array/index.ts
@@ -2,6 +2,7 @@ export { arrCompact } from './arrCompact.js';
 export { arrCount } from './arrCount.js';
 export { arrDifference } from './arrDifference.js';
 export { arrGroupByMaxCount } from './arrGroupByMaxCount.js';
+export { arrIntersection } from './arrIntersection.js';
 export { arrMove } from './arrMove.js';
 export { arrPick } from './arrPick.js';
 export { arrRepeat } from './arrRepeat.js';